ai-firstaid-backend/requirements.txt
# --- Core Server ---
fastapi>=0.115.0
uvicorn>=0.30.0

# --- AI & Machine Learning (PyTorch & Transformers) ---
# Torch 2.3 is safe and stable for both CPU and GPU (CUDA 12+)
torch>=2.3.0
transformers>=4.46.2
accelerate>=0.33.0
bitsandbytes>=0.44.0

# --- RAG Components (Vector Search) ---
faiss-cpu>=1.8.0
sentence-transformers>=3.1.0

# --- General Utilities ---
pydantic>=2.7.0
numpy>=1.26.4
